use useField internally
